backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
sqlalchemy
email-validator
python-multipart